+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintPlugin.kt
@@ -292,6 +292,8 @@ open class KtlintPlugin : Plugin<Project> {
         ruleSets.set(pluginHolder.extension.ruleSets)
         ruleSetsClasspath.setFrom(pluginHolder.ktlintRulesetConfiguration)
         reporters.set(pluginHolder.extension.reporters)
+        customReportersClasspath.setFrom(pluginHolder.ktlintReporterConfiguration)
+        customReporters.set(pluginHolder.extension.customReportersSet)
         android.set(pluginHolder.extension.android)
         enableExperimentalRules.set(pluginHolder.extension.enableExperimentalRules)
 
@@ -382,6 +384,9 @@ open class KtlintPlugin : Plugin<Project> {
 
         val ktlintConfiguration by lazy(LazyThreadSafetyMode.NONE) { createKtlintConfiguration(target, extension) }
         val ktlintRulesetConfiguration = createKtlintRulesetConfiguration(target)
+        val ktlintReporterConfiguration by lazy(LazyThreadSafetyMode.NONE) {
+            createKtlintReporterConfiguration(target, extension)
+        }
     }
 }