+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/reporter/CustomReporter.kt
@@ -1,0 +1,9 @@
+package org.jlleitschuh.gradle.ktlint.reporter
+
+import org.gradle.api.artifacts.Dependency
+
+data class CustomReporter(
+    val reporterId: String,
+    val reporterFileExtension: String,
+    val dependency: Dependency
+)